Add logging to telegraf sensor conf
At this moment telegraf will og to directory /var/log/telegraf which is
default logging directory for telegraf.

I also add code to installation script for changing ownership and group to pi because otherwise telegraf service is not able to log into this directory because of ownership rules. It also prevents telegraf service from logging to different directory.